Add new pet animation
Currently, there is a problem with this. For some avatars the background
flashes -- but only for some and only some frames. It may be none, two
or every other frame that does not have transparent background.

I think that the reason may be the palette GIF is using to put pixels
in place, but I'm not sure. Until this is resolved, it cannot be merged.

I tried to look online, but I have not been succesfull. Maybe there is
some library on pypi that may solve this, I don't know.