Review: Support reviews with arbitrary length
Because Discord allows text messages longer than 2000 characters, it is
possible to create review longer than 2048 characters, which was not
true when the module was written.

This commit fixes it to support reviews with extended length. Regular
embed limitations still apply and it can't be more than 6000 characters,
but that is not possible to reach in this situation.